// src/app/models/project.model.ts
export class Project {
    id: number;
    title: string;
    psk: any;
    lsp: any;
    ilsp: any;
    bns: any;
    cva: any;
    inp: any;
    competencies: any;
    icompetencies: any;
    participans: any;
    myparticipant: any;
    surveys: any;
    survey_questions: any;
    survey_answers: any;
    values: any;
    lsreport: any;
    pskcs: any;

    constructor()  { 
         this.psk = {};
         this.lsp = {};
         this.pskcs = {};
        
    }
    
}